requirements.txt
streamlit
pinecone-client
llama-index
llama-index-vector-stores-pinecone
openai
